requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
pinecone-client==2.2.1
sentence-transformers==2.2.2
groqcloud==0.1.3
langchain==0.0.300
openai==0.27.10
